requirements.txt
# Core dependencies from app.py
requests
pypdf
google-generativeai
chromadb
python-dotenv

# Additional dependency for the Gradio UI
gradio
